add Modal component + remove ImageModal + Contact Modal
fix changes add back image + contact modal

- toggleModal -> handleToggleModal across the code

WIP

- enhanced ModalWrapper to accept children
- more use of self

- improved CallToAction to use new ModalWrapper encompassing
ContactModal

- improved Thumbnail to use new ModalWrapper encompassing ImageModal
- fixing some whitespace

remove modal component in favor of wrapper modal
